requirements.txt
azure-functions
sendgrid
PyNFe
pyodbc
pymongo
